Start CPU side shader impl
This commit has a draft of all CPU stages needed to fill paths. It is missing clips, other paint types than solid colors, and blends. In addition, all of the code exists only as entry points, with no integration glue to wire into a pipeline.

Because the code is untested, no doubt a bunch of bugs exist. Checkpoint commit to prepare for next step, which is beginning to wire the shader stages.

+++ crates/encoding/src/path.rs
@@ -213,7 +213,7 @@ pub struct PathBbox {
 #[repr(C)]
 pub struct Path {
     /// Bounding box in tiles.
-    pub bbox: [f32; 4],
+    pub bbox: [u32; 4],
     /// Offset (in u32s) to tile rectangle.
     pub tiles: u32,
     _padding: [u32; 3],
